requirements.txt
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
faiss-cpu==1.7.4
fastapi==0.95.2
filelock==3.19.1
Flask==3.1.2
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
lightgbm==4.4.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.2
numpy==1.24.3
packaging==25.0
pandas==1.5.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==1.10.24
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.2.2
scipy==1.13.1
sentence-transformers==5.1.1
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.27.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.22.0
wcwidth==0.2.14
Werkzeug==3.1.3
xgboost==2.1.4
zipp==3.23.0